chore(warehouse): remove the deltalite shadow (phase 1 complete)
The shadow verification did its job: it proved deltalite is byte-for-byte
equivalent to the delta-rs MERGE across the fleet on real prod S3, and surfaced
the null-composite-key duplicate bug in MERGE (fixed in #75601). With phase 2
(deltalite doing the real write) in this PR, running the shadow too is redundant.

Removes deltalite_shadow.py (compare engine, seed/upsert/cleanup, sampling +
caps), the DATA_WAREHOUSE_DELTALITE_SHADOW_* settings, the shadow metrics, and
the write_to_deltalake shadow hook. The per-schema flag helper the write path
needs is kept in a focused new module, deltalite_write.py.
