+++ products/warehouse_sources/backend/temporal/data_imports/pipelines/core/delta_table_helper.py
@@ -541,49 +541,6 @@ class DeltaTableHelper:
             )
         return deduped
 
-    async def _maybe_run_deltalite_shadow(
-        self,
-        *,
-        data: pa.Table,
-        primary_keys: list[str],
-        partition_key: str | None,
-        version_before: int,
-        version_after: int | None,
-        commit_metadata: dict[str, str] | None,
-    ) -> None:
-        """Best-effort deltalite shadow verification for this incremental batch.
-
-        Checks the per-schema rollout flag, then re-applies the batch through deltalite into a
-        throwaway copy and compares to the merge result (see ``deltalite_shadow``). Wrapped so that
-        nothing here — flag eval, import, comparison — can ever raise into the real sync. Imported
-        lazily to avoid a circular import (``deltalite_shadow`` reuses ``_purge_s3_prefix``).
-        """
-        try:
-            from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.deltalite_shadow import (
-                is_deltalite_shadow_enabled,
-                run_shadow_comparison,
-            )
-
-            enabled = await database_sync_to_async_pool(is_deltalite_shadow_enabled)(
-                self._job.team_id, str(self._job.schema_id), None
-            )
-            if not enabled:
-                return
-
-            await run_shadow_comparison(
-                uri=await self._get_delta_table_uri(),
-                storage_options=self._get_credentials(),
-                data=data,
-                primary_keys=primary_keys,
-                partition_key=partition_key,
-                version_before=version_before,
-                version_after=version_after,
-                commit_metadata=commit_metadata,
-                logger=self._logger,
-            )
-        except Exception as e:  # noqa: BLE001 - shadow must never affect the sync
-            await self._logger.awarning(f"deltalite shadow wrapper errored (ignored): {e}")
-
     async def _write_via_deltalite(
         self,
         *,
@@ -604,14 +561,14 @@ class DeltaTableHelper:
         """
         if not normalized_primary_keys:
             return False
-        # Lazy imports: keep the heavy pipeline_v3 metrics chain (and deltalite_shadow) off this core
-        # module's import path, which would otherwise risk a circular import.
+        # Lazy import: keep the heavy pipeline_v3 metrics chain off this core module's import path,
+        # which would otherwise risk a circular import.
         from products.warehouse_sources.backend.temporal.data_imports.pipelines.pipeline_v3.load.metrics import (
             DELTALITE_WRITE_TOTAL,
         )
 
         try:
-            from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.deltalite_shadow import (
+            from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.deltalite_write import (
                 is_deltalite_write_enabled,
             )
 
@@ -714,10 +671,6 @@ class DeltaTableHelper:
 
             existing_delta_table = delta_table
 
-            # Captured before the merge so the deltalite shadow (below) can time-travel to the exact
-            # pre-merge state when it re-applies this batch into a throwaway copy.
-            version_before_merge = existing_delta_table.version()
-
             await self._logger.adebug(f"write_to_deltalake: merging...")
 
             # Normalize keys and check the keys actually exist in the dataset
@@ -819,26 +772,6 @@ class DeltaTableHelper:
                     "write_to_deltalake: merge",
                 )
                 await self._logger.adebug(f"Delta Merge Stats: {json.dumps(merge_stats)}")
-
-            # deltalite shadow verification (rollout canary, phase 1). Re-applies this exact batch
-            # through the deltalite streaming upsert into a throwaway copy and compares the result to
-            # the merge above. Best-effort and can never affect the sync. Skipped when deltalite did
-            # the real write (`deltalite_wrote`) — the real table would already be deltalite's output,
-            # so the comparison is trivially true and pointless. Master-switched off by default.
-            if not deltalite_wrote and settings.DATA_WAREHOUSE_DELTALITE_SHADOW_ENABLED:
-                # Version the merge just produced. delta-rs advances the table in place on commit; guard
-                # against it not advancing (fall back to latest) so we never read the pre-merge state.
-                version_after_merge: int | None = existing_delta_table.version()
-                if version_after_merge is None or version_after_merge <= version_before_merge:
-                    version_after_merge = None
-                await self._maybe_run_deltalite_shadow(
-                    data=data,
-                    primary_keys=normalized_primary_keys,
-                    partition_key=PARTITION_KEY if use_partitioning else None,
-                    version_before=version_before_merge,
-                    version_after=version_after_merge,
-                    commit_metadata=commit_metadata,
-                )
         elif (
             write_type == "full_refresh"
             or (write_type == "incremental" and delta_table is None)
