+++ products/warehouse_sources/backend/temporal/data_imports/pipelines/pipeline_v3/load/metrics.py
@@ -47,23 +47,6 @@ DELTA_REPARTITION_SKIP_TOTAL = Counter(
     labelnames=["team_id", "reason"],
 )
 
-# deltalite shadow verification (rollout canary). `outcome` is one of:
-#   match      - deltalite produced byte-identical logical content to the delta-rs MERGE
-#   mismatch   - deltalite disagreed with MERGE (a correctness bug; investigate before rollout)
-#   skipped    - not run (sampled out, over the affected-bytes cap, SCD2, or no affected partitions)
-#   unsupported- deltalite refused the table (deletion vectors / column mapping)
-#   error      - the shadow itself failed (never affects the real sync)
-DELTALITE_SHADOW_TOTAL = Counter(
-    "warehouse_load_deltalite_shadow_total",
-    "deltalite shadow verifications against the delta-rs MERGE, by outcome",
-    labelnames=["outcome"],
-)
-
-DELTALITE_SHADOW_DURATION_SECONDS = Histogram(
-    "warehouse_load_deltalite_shadow_duration_seconds",
-    "Wall-clock time of a deltalite shadow verification (seed + upsert + compare)",
-)
-
 # deltalite real-write path (phase 2). `outcome` is one of:
 #   written    - deltalite performed the incremental merge and committed
 #   fallback   - deltalite was enabled but failed (or refused), so the delta-rs MERGE ran instead
